+++ demis/applications/chart-context_enrichment_service.tf
@@ -35,6 +35,7 @@ module "context_enrichment_service" {
     istio_proxy_resources                              = try(local.ces_resources_overrides.istio_proxy_resources, var.istio_proxy_default_resources)
   })
   istio_values = templatefile(local.ces_template_istio, {
-    namespace = var.target_namespace
+    namespace                = var.target_namespace
+    http_timeout_retry_block = try(module.http_timeouts_retries.service_timeout_retry_definitions[local.ces_name], null)
   })
 }