+++ demis/applications/chart-notification_processing_service.tf
@@ -46,5 +46,6 @@ module "notification_processing_service" {
     support_fhir_api_versions      = var.profile_provisioning_mode_vs_core != null && var.profile_provisioning_mode_vs_core != "dedicated"
     fhir_api_versions              = module.validation_service_core_metadata.current_profile_versions,
     feature_flag_new_api_endpoints = try(var.feature_flags[local.nps_name].FEATURE_FLAG_NEW_API_ENDPOINTS, false)
+    http_timeout_retry_block       = try(module.http_timeouts_retries.service_timeout_retry_definitions[local.nps_name], null)
   })
 }