+++ demis/applications/chart-notification_routing_service.tf
@@ -38,6 +38,7 @@ module "notification_routing_service" {
     istio_proxy_resources                              = try(local.nrs_resources_overrides.istio_proxy_resources, var.istio_proxy_default_resources)
   })
   istio_values = templatefile(local.nrs_template_istio, {
-    namespace = var.target_namespace
+    namespace                = var.target_namespace
+    http_timeout_retry_block = try(module.http_timeouts_retries.service_timeout_retry_definitions[local.nrs_name], null)
   })
 }