+++ demis/applications/chart-pdfgen_service.tf
@@ -37,6 +37,7 @@ module "pdfgen_service" {
     istio_proxy_resources                              = try(local.pdfgen_resources_overrides.istio_proxy_resources, var.istio_proxy_default_resources)
   })
   istio_values = templatefile(local.pdfgen_template_istio, {
-    namespace = var.target_namespace
+    namespace                = var.target_namespace
+    http_timeout_retry_block = try(module.http_timeouts_retries.service_timeout_retry_definitions[local.pdfgen_name], null)
   })
 }