+++ demis/applications/variables.tf
@@ -207,7 +207,6 @@ variable "istio_enabled" {
 }
 
 # Feature Flags
-
 variable "feature_flags" {
   type        = map(map(bool))
   description = "Defines a list of feature flags to be bound in services"
@@ -215,11 +214,30 @@ variable "feature_flags" {
 }
 
 # Operational Flags
-
 variable "config_options" {
   type        = map(map(string))
   description = "Defines a list of ops flags to be bound in services"
   default     = {}
+}
+
+variable "timeout_retry_overrides" {
+  description = "Defines retry and timeout configurations per service. Each definition must include a service name and can optionally include timeout and retry settings."
+  type = list(object({
+    service = string
+    timeout = optional(string)
+    retries = optional(object({
+      enable        = optional(bool)
+      attempts      = optional(number)
+      perTryTimeout = optional(string)
+      retryOn       = optional(string)
+    }))
+  }))
+  default = []
+
+  validation {
+    condition     = length(var.timeout_retry_overrides) > 0 ? alltrue([for conf in var.timeout_retry_overrides : length(conf.service) > 0]) : true
+    error_message = "Service name must not be empty"
+  }
 }
 
 #########################