+++ dmz/applications.tf
@@ -19,9 +19,17 @@ module "dmz_services" {
   config_options                = module.application_flags.service_config_options
   resource_definitions          = module.application_resources.service_resource_definitions
   istio_proxy_default_resources = module.application_resources.istio_proxy_default_resources
+  timeout_retry_overrides       = var.timeout_retry_overrides
   reset_values                  = var.reset_values
   rabbitmq_pvc_config           = var.rabbitmq_pvc_config
   allow_even_rabbitmq_replicas  = var.allow_even_rabbitmq_replicas
+  database_target_host          = var.database_target_host
+  postgres_server_certificate   = var.postgres_server_certificate
+  postgres_root_ca_certificate  = var.postgres_root_ca_certificate
+  postgres_server_key           = var.postgres_server_key
+  database_credentials          = var.database_credentials
+  ars_bulk_upload_hmac_secret   = var.ars_bulk_upload_hmac_secret
+  secure_message_gateway_url    = var.secure_message_gateway_url
 
   depends_on = [module.persistent_volume_claims, module.pull_secrets, module.activate_maintenance_mode.set_maintenance_mode]
 }