+++ dmz/applications/chart-bulk_inbound_service.tf
@@ -11,6 +11,31 @@ locals {
   bulk_inbound_resource_block      = lookup(local.bulk_inbound_resources_overrides, "resource_block", null) != null ? var.resource_definitions[local.bulk_inbound_name].resource_block : null
 
   supported_ars_profile_versions = var.deployment_information["ars-profile-snapshots"].main.profiles
+
+  ars_bulk_ddl_index = try(
+    index(
+      [for cred in var.database_credentials : cred.secret-name],
+      "ars-bulk-ddl-database-secret"
+    ),
+    -1 # Default index if not found
+  )
+
+  ars_bulk_user_index = try(
+    index(
+      [for cred in var.database_credentials : cred.secret-name],
+      "ars-bulk-user-database-secret"
+    ),
+    -1 # Default index if not found
+  )
+
+
+  ars_bulk_purger_index = try(
+    index(
+      [for cred in var.database_credentials : cred.secret-name],
+      "ars-bulk-purger-database-secret"
+    ),
+    -1 # Default index if not found
+  )
 }
 
 module "bulk_inbound_service" {
@@ -32,10 +57,15 @@ module "bulk_inbound_service" {
     debug_enable                                       = var.debug_enabled,
     istio_enable                                       = var.istio_enabled,
     core_hostname                                      = var.core_hostname,
+    context_path                                       = var.context_path,
     feature_flags                                      = try(var.feature_flags[local.bulk_inbound_name], {}),
     config_options                                     = try(var.config_options[local.bulk_inbound_name], {}),
     replica_count                                      = local.bulk_inbound_replicas,
     resource_block                                     = local.bulk_inbound_resource_block
+    ars_bulk_ddl_db_secret_checksum                    = try(kubernetes_secret_v1.database_credentials[local.ars_bulk_ddl_index].metadata[0].annotations["checksum"], ""),
+    ars_bulk_user_db_secret_checksum                   = try(kubernetes_secret_v1.database_credentials[local.ars_bulk_user_index].metadata[0].annotations["checksum"], ""),
+    ars_bulk_purger_db_secret_checksum                 = try(kubernetes_secret_v1.database_credentials[local.ars_bulk_purger_index].metadata[0].annotations["checksum"], ""),
+    ars_bulk_upload_hmac_secret_checksum               = try(var.ars_bulk_upload_hmac_secret.metadata[0].annotations["checksum"], ""),
     feature_flag_new_istio_sidecar_requests_and_limits = try(var.feature_flags[local.bulk_inbound_name].FEATURE_FLAG_NEW_ISTIO_SIDECAR_REQUEST_AND_LIMITS, false)
     istio_proxy_resources                              = try(local.bulk_inbound_resources_overrides.istio_proxy_resources, var.istio_proxy_default_resources)
   })
@@ -46,5 +76,6 @@ module "bulk_inbound_service" {
     fhir_api_versions              = local.supported_ars_profile_versions,
     demis_hostnames                = local.demis_hostnames,
     feature_flag_new_api_endpoints = try(var.feature_flags[local.bulk_inbound_name].FEATURE_FLAG_NEW_API_ENDPOINTS, false)
+    http_timeout_retry_block       = try(module.http_timeouts_retries.service_timeout_retry_definitions[local.bulk_inbound_name], null)
   })
 }