+++ idm/applications/chart-bundid_idp.tf
@@ -41,8 +41,9 @@ module "bundid_idp" {
     istio_proxy_resources                              = try(local.bundid_resources_overrides.istio_proxy_resources, var.istio_proxy_default_resources)
   })
   istio_values = templatefile(local.bundid_template_istio, {
-    namespace       = var.target_namespace,
-    cluster_gateway = var.cluster_gateway,
-    issuer_hostname = local.bundid_idp_hostname
+    namespace                = var.target_namespace,
+    cluster_gateway          = var.cluster_gateway,
+    issuer_hostname          = local.bundid_idp_hostname
+    http_timeout_retry_block = try(module.http_timeouts_retries.service_timeout_retry_definitions[local.bundid_name], null)
   })
 }