+++ modules/http_timeouts_retries/outputs.tf
@@ -1,0 +1,4 @@
+output "service_timeout_retry_definitions" {
+  description = "Map containing all the timeout and retry definitions, grouped by service"
+  value       = local.encoded_map
+}